+++ apps/api/src/app/portfolio/calculator/portfolio-calculator.ts
@@ -53,6 +53,7 @@ import {
   isBefore,
   isWithinInterval,
   min,
+  startOfDay,
   startOfYear,
   subDays
 } from 'date-fns';
@@ -162,8 +163,8 @@ export abstract class PortfolioCalculator {
       subDays(dateOfFirstActivity, 1)
     );
 
-    this.endDate = endDate;
-    this.startDate = startDate;
+    this.endDate = endOfDay(endDate);
+    this.startDate = startOfDay(startDate);
 
     this.computeTransactionPoints();
 
@@ -236,7 +237,7 @@ export abstract class PortfolioCalculator {
     const exchangeRatesByCurrency =
       await this.exchangeRateDataService.getExchangeRatesByCurrency({
         currencies: Array.from(new Set(Object.values(currencies))),
-        endDate: endOfDay(this.endDate),
+        endDate: this.endDate,
         startDate: this.startDate,
         targetCurrency: this.currency
       });
@@ -445,7 +446,6 @@ export abstract class PortfolioCalculator {
         quantity: item.quantity,
         symbol: item.symbol,
         tags: item.tags,
-        transactionCount: item.transactionCount,
         valueInBaseCurrency: new Big(marketPriceInBaseCurrency).mul(
           item.quantity
         )
@@ -1005,8 +1005,7 @@ export abstract class PortfolioCalculator {
             oldAccumulatedSymbol.feeInBaseCurrency.plus(feeInBaseCurrency),
           includeInHoldings: oldAccumulatedSymbol.includeInHoldings,
           quantity: newQuantity,
-          tags: oldAccumulatedSymbol.tags.concat(tags),
-          transactionCount: oldAccumulatedSymbol.transactionCount + 1
+          tags: oldAccumulatedSymbol.tags.concat(tags)
         };
       } else {
         currentTransactionPointItem = {
@@ -1024,8 +1023,7 @@ export abstract class PortfolioCalculator {
           dividend: new Big(0),
           includeInHoldings: INVESTMENT_ACTIVITY_TYPES.includes(type),
           investment: unitPrice.mul(quantity).mul(factor),
-          quantity: quantity.mul(factor),
-          transactionCount: 1
+          quantity: quantity.mul(factor)
         };
       }