+++ apps/api/src/app/subscription/subscription.service.ts
@@ -35,7 +35,7 @@ export class SubscriptionService {
       this.stripe = new Stripe(
         this.configurationService.get('STRIPE_SECRET_KEY'),
         {
-          apiVersion: '2025-12-15.clover'
+          apiVersion: '2026-01-28.clover'
         }
       );
     }
@@ -100,7 +100,6 @@ export class SubscriptionService {
     );
 
     return {
-      sessionId: session.id,
       sessionUrl: session.url
     };
   }